Allow tables to be copied between threads
You can now pass in table objects as the parameters or return values for
a thread.  The contents of the table must themselves be objects that can
be passed into threads.  So (for instance), arbitrarily nested tables of
normal numbers and strings are fine.  But you still can't pass in a
function or a heavy userdata.